buildExploreSection
