+++ src/selectors/__tests__/explore.test.ts
@@ -1,11 +1,12 @@
 // tslint:disable:no-let no-expression-statement
-import { denormalizeSections, buildExploreSection } from '../details/explore';
+import { denormalizeSections } from '../details/explore';
 import { ExploreSection } from '../explore/types';
 import { ExploreSectionMap } from '../../stores/explore';
 import { ExploreTaxonomyId } from '../../stores/taxonomies';
 import { aString } from '../../application/__tests__/helpers/random_test_values';
 import { ExploreSectionBuilder as StoreExploreSectionBuilder } from '../../stores/__tests__/helpers/explore_section_builder';
 import { LocalizedTextBuilder } from '../../stores/__tests__/helpers/locale_helpers';
+import { buildExploreSection } from '../explore/build_explore_section';
 
 const englishLocale = { code: 'en', fallback: 'ar' };
 const theId = aString();
