+++ src/selectors/articles/select_article.ts
@@ -3,12 +3,12 @@ import * as taskDetails from '../details/tasks';
 import { selectLocale } from '../locale';
 import { selectRelatedTasks } from '../tasks';
 import { RouterProps } from '../../application/routing';
-import { buildExploreSection } from '../details/explore';
 import { selectExploreTaxonomy } from '../taxonomies';
 import { selectIconFromExploreTaxonomy } from '../select_icon_from_explore_taxonomy';
 import { toSelectorArticle } from './to_selector_article';
 import { toSelectorArticleList } from './to_selector_article_list';
 import { Article } from './types';
+import { buildExploreSection } from '../explore/build_explore_section';
 
 export const selectArticle = (store: Store, routerProps: RouterProps): Article => {
     const locale = selectLocale(store);
