+++ src/selectors/explore/select_explore_section.ts
@@ -1,10 +1,10 @@
 import { Store } from '../../stores';
-import * as details from '../details/explore';
 import { selectLocale } from '../locale';
 import { selectExploreTaxonomy } from '../taxonomies';
 import { RouterProps } from '../../application/routing';
 import { selectIconFromExploreTaxonomy } from '../select_icon_from_explore_taxonomy';
 import { ExploreSection } from './types';
+import { buildExploreSection } from './build_explore_section';
 
 export const selectExploreSection = (store: Store, routerProps: RouterProps): ExploreSection => {
     const locale = selectLocale(store);
@@ -13,5 +13,5 @@ export const selectExploreSection = (store: Store, routerProps: RouterProps): Ex
     const theSection = sections[id];
     const exploreTaxonomy = selectExploreTaxonomy(store);
     const icon = selectIconFromExploreTaxonomy(theSection.taxonomyTerms, exploreTaxonomy);
-    return details.buildExploreSection(locale, theSection, icon);
+    return buildExploreSection(locale, theSection, icon);
 };
